Improve font hashing scheme in FontLoaderWin to decrease duplicates in font cache (reduces memory, especially with large fonts).

+++ src/win/FontLoaderWin.h
@@ -14,7 +14,7 @@ public:
   virtual String16 fallback_font_for_characters(const String16& characters, int weight, bool italic, float size) const override;
   virtual Ref<Buffer> Load(const String16& family, int weight, bool italic, float size) override;
 protected:
-  std::map<uint64_t, RefPtr<Buffer>> fonts_;
+  std::map<uint32_t, RefPtr<Buffer>> fonts_;
 };
 
 }  // namespace ultralight
